+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/TreeDiff.hs
@@ -257,3 +257,5 @@ instance ToExpr (PParams era) => ToExpr (CertEnv era)
 instance ToExpr (PParams era) => ToExpr (ConwayDelegEnv era)
 
 instance ToExpr (PParamsHKD StrictMaybe era) => ToExpr (RatifySignal era)
+
+instance ToExpr (PParamsHKD StrictMaybe era) => ToExpr (EnactSignal era)