+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/Spec/Conway.hs
@@ -19,6 +19,7 @@ spec = describe "Conway conformance tests" $ do
   prop "CERT" $ conformsToImpl @"CERT" @ConwayFn @Conway
   xprop "RATIFY" $ conformsToImpl @"RATIFY" @ConwayFn @Conway
   prop "GOVCERT" $ conformsToImpl @"GOVCERT" @ConwayFn @Conway
+  xprop "ENACT" $ conformsToImpl @"ENACT" @ConwayFn @Conway
   describe "Generators" $ do
     let
       genEnv = do